Split world.cpp into logical subsystem homes
Decompose the monolithic world.cpp into focused subsystem files, leaving
world.cpp as a cohesive WorldInfo module (core API + world-loader globals
+ players globals + world-module helpers):

- render.cpp: nu3d / NuGScn render + camera + gfx upload
- terrain.cpp: terrain / debris / grass / bridge / particles
- sfx.cpp: sound / SFX bits / quiet tables
- gizmo_sys.cpp: gizmo / minikit / collection stubs
- ai_sys.cpp: AI system + API object system
- game_object.cpp: game / api objects
- cutscene.cpp: cutscenes + character scenes
- levelstubs.cpp: remaining WorldInfo_Load dependency stubs

New src/legoapi/world_shared.h provides cross-file prototypes and shared
globals, preserving extern "C" linkage so mangled names still match the
original libTTapp.so. Register the new sources in CMakeLists.txt.
