+++ src/legoapi/world.cpp
@@ -1,4 +1,5 @@
 #include "legoapi/world.h"
+#include "legoapi/world_shared.h"
 
 #include <stdio.h>
 #include <string.h>
@@ -20,75 +21,11 @@
 #include "nu2api/nufile/nufile.h"
 #include "nu2api/nufile/nufpar.h"
 
-extern "C" {
-    void NuDisplaySceneRndr(void *scene) {
-        (void)scene;
-    }
-
-    void NuPortalVisibility(NUGSCN *scene) {
-        (void)scene;
-    }
-    void NuGScnRndr3(NUGSCN *scene) {
-        NuDisplaySceneRndr(scene->display_list);
-    }
-}
-
-// Debris and camera globals — accessed from DebrisSetThinningLevel etc.
-f32 debris_thinning_level;
-i32 forced_debris_thinning;
-i32 debris_detail_level;
-f32 CameraZoom;
-
-// Graphics loading flags
-i32 RemoveDirectionalMaps = 0;
-i32 RemoveNormalMaps = 0;
-
-extern "C" void DebrisSetThinningLevel(f32 level) {
-    debris_thinning_level = level < 1.0f ? 1.0f : level;
-}
-extern "C" void DebrisSetForcedThinning(i32 forced) {
-    forced_debris_thinning = forced;
-}
-extern "C" void DebrisSetDetailLevel(i32 level) {
-    debris_detail_level = level;
-}
-void SetCameraZoom(f32 zoom) {
-    CameraZoom = zoom;
-}
-i32 ActionFromQuiet(i32 idx) {
-    static i16 ActionPairTab[14] = {-1};
-    if (idx != -1) {
-        i16 *pair = ActionPairTab;
-        while (*pair != -1) {
-            if (*pair == idx) {
-                return pair[1];
-            }
-            pair += 14;
-        }
-    }
-    return -1;
-}
-i32 AmbientFromQuiet(i32 idx) {
-    static i16 AmbientPairTab[2] = {-1};
-    if (idx != -1) {
-        i16 *pair = AmbientPairTab;
-        while (*pair != -1) {
-            if (*pair == idx) {
-                return pair[1];
-            }
-            pair += 2;
-        }
-    }
-    return -1;
-}
-
 // Globals shared across world loading — defined here until moved to globals.cpp
 TIMER *LevelTimer = NULL;
 i32 abort_load = 0;
 char *ConfigBuffer = NULL;
 i32 numtl_force_mipmode = 0;
-char *debris_name = NULL;
-void *MiniKitCollection = NULL;
 i32 GAMEDEMO = 0;
 void *big_icon_scene = NULL;
 void *area_scene = NULL;
@@ -98,78 +35,6 @@ void *RETAKED_LDATA = NULL;
 void *CREDITS_LDATA = NULL;
 u32 Text_Language = 0;
 nufpcomjmp_s *LevelConfigKeywords_AfterLoad = NULL;
-i32 Grass_Available = 1;
-i32 PDEBCOUNT = 0;
-i32 CharScene_Area = 0;
-void *PDebNameList = NULL;
-
-// Stub implementations for functions not yet implemented in their respective modules.
-// NOTE: Functions that have C++ linkage in the original libTTapp.so are kept
-// outside extern "C" so their mangled names match.  Functions with true C
-// linkage in the original are individually marked `extern "C"`.
-
-extern "C" {
-    void ResetSounds(void) {
-    }
-} // extern "C"
-
-// --- C++ linkage stubs (original uses C++ mangling) ---
-
-void SetLevelSfxBits(WORLDINFO *world) {
-    (void)world;
-}
-void ResetLevSfx(WORLDINFO *world) {
-    // SFX bit array and counter in the unknown_4670[] filler region.
-    // TODO: these offsets must be replaced with typed struct fields.
-    //   0x4720 → unknown_4670[0xb0]: SFX bit array (0x400 bytes, stride 0x10)
-    //   0x4b14 → unknown_4670[0x4a4]: SFX counter
-    i16 *sfx = (i16 *)&world->unknown_4670[0xb0];
-    for (i32 i = 0; i < 0x40; i++) {
-        sfx[i] = -1;
-    }
-    *(i32 *)&world->unknown_4670[0x4a4] = 0;
-}
-
-extern "C" void noterraininit(void) {
-}
-extern "C" void TerrainSetCur(void *terrain) {
-    (void)terrain;
-}
-extern "C" void TerrSetPlatScanDist(f32 dist) {
-    (void)dist;
-}
-extern "C" void TerrainPlatformOldUpdate(void) {
-}
-extern "C" void TerrainPlatformNewUpdate(void) {
-}
-extern "C" void TerrainSetWallDeflectYScale(f32 scale) {
-    (void)scale;
-}
-
-// --- More C++ linkage stubs ---
-
-void LevObj_FixUpPlatIDs(WORLDINFO *world) {
-    (void)world;
-}
-extern "C" void NuRndrInitWorld(void); // Defined in nurndr_android.c
-
-extern "C" void NuGScnUpdate(NUGSCN *gscn, i32 param) {
-    (void)gscn;
-    (void)param;
-}
-
-void Doors_Init(WORLDINFO *world) {
-    (void)world;
-}
-
-
-
-void ClearGameObjects(void *api_object_sys) {
-    (void)api_object_sys;
-}
-
-
-
 
 // --- Players globals ---
 // NOTE: `WORLD` (WORLDINFO) field offsets in world.h are an approximation; the
@@ -204,688 +69,27 @@ u16 LEGOOBJ_DEFAULTLASTCOIN = 50;
 
 PLAYERDATA *apicharsys;
 
-// --- Helper stubs (implemented separately) ---
-
-GameObject_s *AddGameObject(i32 id) {
-    (void)id;
-    UNIMPLEMENTED();
-    return NULL;
-}
-i32 InitCreature(GameObject_s *obj, i32 id, i32 param) {
-    (void)obj;
-    (void)id;
-    (void)param;
-    UNIMPLEMENTED();
-    return 0;
-}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 TORPEDOPACKET TorpedoPackets[16];
 
+// --- World-module helpers (kept with the WorldInfo API) ---
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-// ---
-
-extern "C" void rtlResetDynamic(void) {
-}
-extern "C" void SetPartRTLSet(i32 rtl_set) {
-    (void)rtl_set;
-}
-
-void InitGameObjectLights(void) {
-}
-void NewMenu(i32 menu_id, i32 menu_y, i32 param3) {
-    (void)menu_id;
-    (void)menu_y;
-    (void)param3;
-}
-
-// Global variables needed by loading functions
-
-// Stubs for WorldInfo_Load dependencies
-void *TerrainInitEx(i32 param1, void *buf, void *buf_end, i32 param2, char *path, void *gscn, i32 param3, u32 param4,
-                    u32 param5, u32 param6) {
-    (void)param1;
-    (void)buf;
-    (void)buf_end;
-    (void)param2;
-    (void)path;
-    (void)gscn;
-    (void)param3;
-    (void)param4;
-    (void)param5;
-    (void)param6;
-    return NULL;
-}
-void *InitPartDebris(void *buf, void *buf_end, i32 param1, i32 param2, void *param3, i32 page) {
-    (void)buf;
-    (void)buf_end;
-    (void)param1;
-    (void)param2;
-    (void)param3;
-    (void)page;
-    return NULL;
-}
-void LoadTerrainFile(WORLDINFO *world) {
-    char path[256];
-    *(i32 *)&world->unknown_0140[0x281c] = 0;
-    if ((world->current_level->flags & LEVEL_UNKNOWN_FLAG_8) != 0) {
-        NuStrCpy(path, world->config_file);
-        LEVELDATA *level = world->current_level;
-        if (level == (LEVELDATA *)PLATFORM_LDATA) {
-            NuStrCpy(path, "levels\\episode_v\\cloudcityescape\\cloudcityescape_b\\cloudcityescape_b");
-            level = world->current_level;
-        }
-        world->giz_buffer.addr = ALIGN(world->giz_buffer.addr, 4);
-        void *terrain = TerrainInitEx(*(i32 *)&world->unknown_0140[0x011c], &world->giz_buffer, &world->unknown_0108, 0,
-                                      path, world->current_gscn, 0, (u32)(u16)level->max_ter_groups,
-                                      (u32)(u16)level->max_ter_groups, (u32)(u16)level->max_ter_platforms);
-        *(void **)&world->unknown_0140[0x281c] = terrain;
-    }
-}
-void LoadGrassFile(WORLDINFO *world) {
-    char path[268];
-    *(i32 *)&world->unknown_0140[0x2964] = -1;
-    Grass_Available = 1;
-    char *config = world->config_file;
-
-    if (g_isLowEndDevice != 0) {
-        char *found = NuStrIStr(config, "Gungan");
-        if (found == NULL)
-            found = NuStrIStr(config, "SpeederChase");
-        if (found == NULL)
-            found = NuStrIStr(config, "EndorBattle");
-        if (found == NULL)
-            found = NuStrIStr(config, "Retake");
-
-        if (found != NULL) {
-            Grass_Available = 0;
-        } else {
-            Grass_Available = 1;
-        }
-    }
-
-    if (g_isMidRangeDevice != 0 && NuStrIStr(config, "GunGan_A") != NULL) {
-        Grass_Available = 0;
-        return;
-    }
-
-    char *found = NuStrIStr(config, "SpeederChase");
-    if (found != NULL) {
-        Grass_Available = 0;
-    }
-
-    if (Grass_Available != 0) {
-        sprintf(path, "%s.gra", config);
-        if (NuFileExists(path)) {
-            i32 page = edgraLoadPage(path, world->current_gscn, *(i32 *)&world->unknown_0140[0x281c],
-                                     &world->giz_buffer, &world->unknown_0108);
-            *(i32 *)&world->unknown_0140[0x2964] = page;
-        }
-    }
-}
-void LoadBridgeFile(WORLDINFO *world) {
-    char path[256];
-    *(i32 *)&world->unknown_0140[0x2968] = -1;
-    sprintf(path, "%s.bri", world->config_file);
-    if (NuFileExists(path)) {
-        i32 page = edbriLoadPage(path, world->current_gscn);
-        *(i32 *)&world->unknown_0140[0x2968] = page;
-    }
-}
-void LoadPartFile(WORLDINFO *world) {
-    char path[256];
-    *(i32 *)&world->unknown_0140[0x295c] = -1;
-    edpartSetParticlePage(*(i32 *)&world->unknown_0140[0x2958]);
-
-    if ((world->current_level->flags & (LEVEL_OUTRO | LEVEL_MIDTRO | LEVEL_INTRO)) == 0) {
-        sprintf(path, "%s.par", world->config_file);
-        i32 page = -1;
-        if (NuFileExists(path)) {
-            page = edpartLoadPage(path, 1, world->current_gscn);
-            *(i32 *)&world->unknown_0140[0x295c] = page;
-        }
-        void *partDebrisSys =
-            InitPartDebris(&world->giz_buffer, &world->unknown_0108, 0x40, PDEBCOUNT, PDebNameList, page);
-        *(void **)&world->unknown_0140[0x2960] = partDebrisSys;
-    }
-}
-void Particles_Load(WORLDINFO *world, char **debris_name, i32 count, i32 flags) {
-    (void)world;
-    (void)debris_name;
-    (void)count;
-    (void)flags;
-}
-void *CreateGizmoSys(void *world, void *buf, void *buf_end) {
-    (void)world;
-    (void)buf;
-    (void)buf_end;
-    return NULL;
-}
-void LoadGizmoSys(void *gizmo_sys, void *world, char *config_file) {
-    (void)gizmo_sys;
-    (void)world;
-    (void)config_file;
-}
-void LoadEditorSplines(char *path, void *buf, void *buf_end) {
-    (void)path;
-    (void)buf;
-    (void)buf_end;
-}
-void GizmoBlowupResetNameTable(void) {
-}
-void Hub_LoadAndFixUpMiniKits(WORLDINFO *world, void *buf, void *buf_end) {
-    (void)world;
-    (void)buf;
-    (void)buf_end;
-}
-void MiniKit_Load(void *minikit, i32 id, void *buf, void *buf_end, void *param) {
-    (void)minikit;
-    (void)id;
-    (void)buf;
-    (void)buf_end;
-    (void)param;
-}
-void MiniKit_InitPieces(void *minikit, i32 count, void *buf) {
-    (void)minikit;
-    (void)count;
-    (void)buf;
-}
-struct CUTINFO_LOAD {
-    void *cutscene;
-    void *instance;
-    char data[0x13c];
-};
-
-struct CUTSYS_LOAD {
-    void *entries;
-    i32 count;
-    void *end;
-};
-
-i32 CUTCOUNT = 0;
-CUTINFO_LOAD *CutList = NULL;
-i32 ACTIVECUTCOUNT = 0;
-i32 CS_area = 0;
-CUTSYS_LOAD *CS_cutsys = NULL;
-WORLDINFO *CS_worldinfo = NULL;
-f32 CutSceneScale = 1.0f;
-i32 CUTCAM = 0;
-extern i32 InStory(void) {
-}
-
-// These routines are implemented by the cutscene subsystem.  Keep the
-// declarations here rather than hiding the calls behind local no-ops:
-// CutScenes_Load is part of the original loader and its ABI calls these
-// entry points directly.
-extern "C" {
-    extern void *NuGCutSceneLoad(char *name, NUGSCN *gscn1, NUGSCN *gscn2, i32 flags);
-    extern void NuGCutSceneFixUp(void *cutscene, char *name, i32 flags, VARIPTR *end);
-    extern void NuGCutSceneFixUpExtra(void *cutscene, i32 area);
-    extern void *instNuGCutSceneCreate(void *cutscene, i32 flags, i32 param, char *name);
-}
-
-__attribute__((noinline)) static void CutScene_Configure_Load(CUTINFO_LOAD *cut, char *name, VARIPTR *buf,
-                                                              VARIPTR *buf_end) {
-    (void)buf;
-    (void)buf_end;
-    memset(cut, 0, sizeof(*cut));
-    NuStrCpy((char *)cut->data, name);
-}
-
-void *CutScenes_Load(char *config, NUGSCN *gscn1, NUGSCN *gscn2, i32 param1, VARIPTR *buf, VARIPTR *buf_end, i32 param2,
-                     i32 param3, WORLDINFO *world) {
-    NUFPAR *fp;
-    CUTSYS_LOAD *sys;
-    void *initial;
-    CUTINFO_LOAD *cut;
-    char name[128];
-    char path[128];
-    char full_path[128];
-    CUTINFO_LOAD *entries[32];
-
-    if (CutList != NULL && CUTCOUNT > 0) {
-        sys = (CUTSYS_LOAD *)ALIGN(buf->addr, 4);
-        sys->entries = sys + 1;
-        sys->count = CUTCOUNT;
-        CS_area = param2;
-        CS_worldinfo = world;
-        CS_cutsys = sys;
-        buf->void_ptr = (char *)sys->entries + CUTCOUNT * 0x144;
-        return sys;
-    }
-    if (config == NULL) {
-        return NULL;
-    }
-
-    fp = NuFParCreateMem((char *)"cutscenes", config, 0xffff);
-    if (fp == NULL) {
-        return NULL;
-    }
-
-    initial = buf->void_ptr;
-    sys = (CUTSYS_LOAD *)ALIGN((usize)initial, 4);
-    sys->entries = sys + 1;
-    sys->count = 0;
-    sys->end = (char *)sys + 0xc;
-    CS_area = param2;
-    CS_worldinfo = world;
-    CS_cutsys = sys;
-    buf->void_ptr = (char *)sys->entries + ((CHARCOUNT + 0x1f) >> 5) * 4;
-
-    while (NuFParGetLine(fp) != 0) {
-        if (NuFParGetWord(fp) == 0 || NuStrICmp(fp->word_buf, "cutscene") != 0 || sys->count > 0x1f ||
-            NuFParGetWord(fp) == 0) {
-            continue;
-        }
-
-        cut = (CUTINFO_LOAD *)buf->void_ptr;
-        entries[sys->count] = cut;
-        buf->void_ptr = (char *)cut + 0x144;
-        NuStrCpy(name, fp->word_buf);
-        for (char *lower = name; *lower != '\0'; ++lower) {
-            *lower = (char)NuToLower((u8)*lower);
-        }
-        i32 len = NuStrLen(name);
-        while (len > 0 && name[len - 1] != '.') {
-            --len;
-        }
-        if (len > 0) {
-            name[len - 1] = '\0';
-        }
-
-        if (name[0] == 'c' && name[1] == 'u' && name[2] == 't' && name[3] == '\\') {
-            NuStrCpy(path, name);
-        } else {
-            NuStrCpy(path, "cut\\");
-            NuStrCat(path, name);
-            NuStrCat(path, ".cut");
-        }
-        NuStrCpy(full_path, "cutscenes");
-        NuStrCat(full_path, "\\");
-        NuStrCat(full_path, path);
-        CutScene_Configure_Load(cut, full_path, buf, buf_end);
-
-        if ((cut->data[0x51] & 8) == 0 && !InStory()) {
-            continue;
-        }
-        buf->void_ptr = (char *)ALIGN(buf->addr, 0x40);
-        NuStrCpy(full_path, "cutscenes");
-        NuStrCat(full_path, "\\");
-        NuStrCat(full_path, path);
-        cut->cutscene = NuGCutSceneLoad(full_path, gscn1, gscn2, 0);
-        if (cut->cutscene == NULL) {
-            continue;
-        }
-        NuGCutSceneFixUp(cut->cutscene, name, 0, buf_end);
-        NuGCutSceneFixUpExtra(cut->cutscene, param1);
-        cut->instance = instNuGCutSceneCreate(cut->cutscene, 0, 0, name);
-        if (cut->instance != NULL) {
-            *(f32 *)((char *)cut->instance + 0x98) = *(f32 *)((char *)cut + 0x5c) * CutSceneScale;
-            buf->void_ptr = (char *)ALIGN(buf->addr, 0x10);
-        }
-        sys->count++;
-    }
-    NuFParDestroy(fp);
-    if (sys->count == 0) {
-        buf->void_ptr = initial;
-        return NULL;
-    }
-    buf->void_ptr = initial;
-    memmove(initial, entries, sys->count * sizeof(void *));
-    buf->void_ptr = (void *)ALIGN((usize)initial + sys->count * 4, 4);
-    sys->entries = initial;
-    return sys;
-}
-void CharScenes_LevelLoad(WORLDINFO *world) {
-    if (CHARCOUNT <= 0) {
-        return;
-    }
-
-    for (int i = 0; i < CHARCOUNT; i++) {
-        void **entry = (void **)(*(i32 *)&world->char_scene_info.minikit.field_0x18 + i * 0x10);
-        *entry = NULL;
-
-        // Check if we should load this character scene
-        if ((CharScene_Area == 0 || *(i32 *)(CharScene_Area + i * 0x10) == 0) && (CDataList[i].flags & 1) != 0 &&
-            world->cutscene_sys != NULL) {
-            // Check if this character is in a cutscene
-            u32 *cutscene_flags = *(u32 **)((char *)world->cutscene_sys + 8);
-            u32 flag = (cutscene_flags[i >> 5] >> (i & 0x1f)) & 1;
-            if (flag != 0) {
-                // Load the character scene
-                char path[136];
-                VARIPTR buf_end = world->unknown_0108;
-                sprintf(path, "chars\\%s\\%s.gsc", CDataList[i].dir, CDataList[i].file);
-                NUGSCN *scene = (NUGSCN *)NuGScnRead(&world->giz_buffer, buf_end, path);
-                *entry = scene;
-                if (scene != NULL) {
-                    NuSpecialFind(scene, (void **)(entry + 1), CDataList[i].file);
-                }
-            }
-        }
-    }
-}
-
-// --- Extern "C" block: functions with confirmed C linkage in original libTTapp.so ---
-extern "C" {
-
-    void *NuGCutSceneLoad(char *name, NUGSCN *gscn1, NUGSCN *gscn2, i32 flags) {
-        (void)name;
-        (void)gscn1;
-        (void)gscn2;
-        (void)flags;
-        return NULL;
-    }
-    void NuGCutSceneFixUp(void *cutscene, char *name, i32 flags, VARIPTR *end) {
-        (void)cutscene;
-        (void)name;
-        (void)flags;
-        (void)end;
-    }
-    void NuGCutSceneFixUpExtra(void *cutscene, i32 area) {
-        (void)cutscene;
-        (void)area;
-    }
-    void *instNuGCutSceneCreate(void *cutscene, i32 flags, i32 param, char *name) {
-        (void)cutscene;
-        (void)flags;
-        (void)param;
-        (void)name;
-        return NULL;
-    }
-} // extern "C"
-
-// --- C++ linkage stubs (original uses C++ mangling for these) ---
-
-void NuSpecialFind(NUGSCN *scene, void **dest, char *name) {
-    (void)scene;
-    (void)dest;
-    (void)name;
-}
-void LevelSplines_InitForLevel(WORLDINFO *world) {
-    (void)world;
-}
-void LevelObjects_InitForLevel(WORLDINFO *world) {
-    (void)world;
-}
-void BoltTypes_Init(WORLDINFO *world) {
-    (void)world;
-}
-void BoltTypes_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void EquivalentObjects_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void Teleports_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void Doors_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void Faders_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void CharPlatforms_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void Grabber_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void Pulses_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void TrafficAnimSys_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void SpecialMiniKits_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void GizForceSFX_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-
-// --- Extern "C": SockSys functions have C linkage in original ---
-extern "C" {
-    void SockSys_Configure(void *sock_sys, char *config, i32 param, void *buf, void *buf_end, void *gscn) {
-        (void)sock_sys;
-        (void)config;
-        (void)param;
-        (void)buf;
-        (void)buf_end;
-        (void)gscn;
-    }
-    void SockSys_GenerateData(void *sock_sys, void *buf, void *buf_end) {
-        (void)sock_sys;
-        (void)buf;
-        (void)buf_end;
-    }
-} // extern "C"
-
-// --- More C++ linkage stubs ---
-
-void CharacterMiniKits_Load(void *collection, WORLDINFO *world, void *buf, void *buf_end) {
-    (void)collection;
-    (void)world;
-    (void)buf;
-    (void)buf_end;
-}
-void *AISysLoad(void *buf, void *buf_end, i32 size, void *gscn, char *dir, char *name, char *param) {
-    (void)buf;
-    (void)buf_end;
-    (void)size;
-    (void)gscn;
-    (void)dir;
-    (void)name;
-    (void)param;
-    return NULL;
-}
-void *AIPathCnxControlSysCreate(void *buf, void *buf_end, i32 count) {
-    (void)buf;
-    (void)buf_end;
-    (void)count;
-    return NULL;
-}
-void *AIPathCnxHelperSysCreate(void *buf, void *buf_end, i32 count) {
-    (void)buf;
-    (void)buf_end;
-    (void)count;
-    return NULL;
-}
-void *AITriggerSetSysCreate(void *buf, void *buf_end) {
-    (void)buf;
-    (void)buf_end;
-    return NULL;
-}
-void GameAIScriptAddLevelSfx(WORLDINFO *world, void *scripts) {
-    (void)world;
-    (void)scripts;
-}
-void *CreateClimbObjectSys(void *buf, void *buf_end, i32 count) {
-    (void)buf;
-    (void)buf_end;
-    (void)count;
-    return NULL;
-}
-void *APIObjectSysInit(i32 size, void *buf, void *buf_end) {
-    (void)size;
-    (void)buf;
-    (void)buf_end;
-    return NULL;
-}
-void LoadLights(WORLDINFO *world, char *path) {
-    (void)world;
-    (void)path;
-}
-
-// --- Extern "C": rtl functions have C linkage in original ---
-extern "C" {
-    i32 rtlFindByUserId(i32 rtl_set, i32 user_id) {
-        (void)rtl_set;
-        (void)user_id;
-        return -1;
-    }
-    void rtlGetDirection(i32 rtl_set, i32 id, void **out) {
-        (void)rtl_set;
-        (void)id;
-        (void)out;
-    }
-} // extern "C"
-
-// --- More C++ linkage stubs ---
-
-void RippleEffects_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void PortalDoors_Configure(WORLDINFO *world, char *config) {
-    (void)world;
-    (void)config;
-}
-void GizmoSysAddGizmos(void *gizmo_sys, void *giz_flow, void *world) {
-    (void)gizmo_sys;
-    (void)giz_flow;
-    (void)world;
-}
-void *LoadGizFlow(void *world, void *gizmo_sys, char *path, void *buf, void *buf_end) {
-    (void)world;
-    (void)gizmo_sys;
-    (void)path;
-    (void)buf;
-    (void)buf_end;
-    return NULL;
-}
-void InitSpecialSfx(WORLDINFO *world) {
-    (void)world;
-}
-void LoadSpecialSfxFile(WORLDINFO *world) {
-    (void)world;
+void SetAreaPickupGravity(i32 area, i32 level) {
+    (void)area;
+    (void)level;
 }
 void WorldInfo_Dump(WORLDINFO *world) {
     (void)world;
 }
-void StoreSceneProgress(void *gscn, void *progress, i32 param) {
+void StoreSceneProgress(NUGSCN *gscn, SCENEPROGRESS_s *progress, i32 param) {
     (void)gscn;
     (void)progress;
     (void)param;
 }
-void SaveSceneObjectAnimTFactors(void *gscn) {
+void SaveSceneObjectAnimTFactors(NUGSCN *gscn) {
     (void)gscn;
 }
 void CalculateWorldSize(WORLDINFO *world) {
     (void)world;
-}
-void *GameAnimSys_Create(void *buf, void *buf_end) {
-    (void)buf;
-    (void)buf_end;
-    return NULL;
-}
-void *GameAntnode_CreateSys(WORLDINFO *world, void *buf, void *buf_end, i32 count) {
-    (void)world;
-    (void)buf;
-    (void)buf_end;
-    (void)count;
-    return NULL;
-}
-
-// --- Extern "C": NuGScn functions have C linkage in original ---
-extern "C" {
-    i32 NuGScnUploadGfxDataFromFilePS(NUFILE file, VARIPTR *buf, VARIPTR buf_end) {
-        (void)file;
-        (void)buf;
-        (void)buf_end;
-        return 0;
-    }
-
-    void *NuGScnRead(VARIPTR *buf, VARIPTR buf_end, char *path) {
-        extern NUGSCN *NuReadGraphicsData(VARIPTR *, VARIPTR *, char *, NUGSCN *);
-        RemoveDirectionalMaps = 1;
-        RemoveNormalMaps = 1;
-        NUGSCN *scene = NuReadGraphicsData(buf, &buf_end, path, NULL);
-        RemoveNormalMaps = 0;
-        RemoveDirectionalMaps = 0;
-        return scene;
-    }
-    void NuGScnReadFromMemory(NUGSCN *scene) {
-        extern NUGSCN *NuReadGraphicsData(VARIPTR *, VARIPTR *, char *, NUGSCN *);
-        NuReadGraphicsData(NULL, NULL, NULL, scene);
-    }
-    void NuGHGFixup(NUGSCN *scene) {
-        NuGScnReadFromMemory(scene);
-    }
-
-    NUGSCN *NuReadGraphicsData(VARIPTR *buf, VARIPTR *buf_end, char *path, NUGSCN *scene) {
-        if (scene != NULL) {
-            return scene;
-        }
-
-        char converted_path[1033];
-        NuFileExtConvert(converted_path, path, 0x400);
-        NUFILE file = NuFileOpen(converted_path, NUFILE_READ);
-        if (file == 0) {
-            return NULL;
-        }
-
-        i32 file_size = (i32)NuFileOpenSize(file);
-        buf->addr = ALIGN(buf->addr, 0x20);
-        i32 uploaded = NuGScnUploadGfxDataFromFilePS(file, buf, *buf_end);
-        NUGSCN *loaded = (NUGSCN *)ALIGN(buf->addr, 0x20);
-        buf->addr = (usize)((char *)loaded + file_size - uploaded);
-        NuFileRead(file, loaded, file_size - uploaded);
-        NuFileClose(file);
-        return loaded;
-    }
-} // extern "C"
-void SetAreaPickupGravity(i32 area, i32 level) {
-    (void)area;
-    (void)level;
 }
 
 // Placeholder for LEVEL_PROGRESS structure
@@ -1037,7 +241,7 @@ void WorldInfo_Init(WORLDINFO *world) {
 
     Doors_Init(world);
     Players_InitPositions(world);
-    ClearGameObjects(world->ai_sys);
+    ClearGameObjects((APIOBJECTSYS_s *)world->ai_sys);
     PlayerItemTypes_Reset(world);
     Players_Init();
     rtlResetDynamic();
@@ -1148,7 +352,7 @@ void WorldInfo_Load(WORLDINFO *world) {
         world->current_gscn = (NUGSCN *)NuGScnRead(&world->giz_buffer, world->unknown_0108, buf);
         numtl_force_mipmode = 0;
 
-        StoreSceneProgress(world->current_gscn, (void *)&world->filler0[0x15c], 1);
+        StoreSceneProgress(world->current_gscn, (SCENEPROGRESS_s *)&world->filler0[0x15c], 1);
         SaveSceneObjectAnimTFactors(world->current_gscn);
 
         if (world->current_gscn != NULL) {
@@ -1195,7 +399,7 @@ void WorldInfo_Load(WORLDINFO *world) {
     LoadEditorSplines(buf, &world->giz_buffer, &world->unknown_0108);
 
     GizmoBlowupResetNameTable();
-    LoadGizmoSys(world->gizmo_sys, world, world->config_file);
+    LoadGizmoSys((GIZMOSYS_s *)world->gizmo_sys, world, world->config_file);
     if (abort_load != 0)
         goto abort;
 
@@ -1213,7 +417,7 @@ void WorldInfo_Load(WORLDINFO *world) {
         MiniKit_Load(&world->char_scene_info.minikit, (i32)(i16)world->area->minikit_id, &world->giz_buffer,
                      &world->unknown_0108, NULL);
         if (world->char_scene_info.minikit.gscn != NULL) {
-            MiniKit_InitPieces(&world->char_scene_info.minikit, 10, &world->giz_buffer);
+            MiniKit_InitPieces(&world->char_scene_info.minikit, 10, &world->giz_buffer, &world->unknown_0108);
         }
         if (abort_load != 0)
             goto abort;
@@ -1284,7 +488,7 @@ after_area:
     // Terrain/grass/bridge loading
     if (world->area != NULL && (world->area->flags & 5) == 5 &&
         (level->flags & (LEVEL_STATUS | LEVEL_OUTRO | LEVEL_MIDTRO | LEVEL_INTRO)) == 0) {
-        CharacterMiniKits_Load(MiniKitCollection, world, &world->giz_buffer, &world->unknown_0108);
+        CharacterMiniKits_Load((COLLECTION_s *)MiniKitCollection, world, &world->giz_buffer, &world->unknown_0108);
         if (abort_load != 0)
             goto abort;
     } else {
@@ -1360,14 +564,14 @@ after_area:
             goto abort;
     }
 
-    GizmoSysAddGizmos(world->gizmo_sys, *(void **)&world->unknown_0140[0x298c], world);
+    GizmoSysAddGizmos((GIZMOSYS_s *)world->gizmo_sys, (GIZFLOW_s *)*(void **)&world->unknown_0140[0x298c], world);
     if (abort_load != 0)
         goto abort;
 
     // Load gizmo flow
     NuStrCpy(buf, world->config_file);
     NuStrCat(buf, ".git");
-    world->giz_flow = LoadGizFlow(world, world->gizmo_sys, buf, &world->giz_buffer, &world->unknown_0108);
+    world->giz_flow = LoadGizFlow(world, (GIZMOSYS_s *)world->gizmo_sys, buf, &world->giz_buffer, &world->unknown_0108);
     if (abort_load != 0)
         goto abort;
